feat: update character model to support multi-level stat tracking and skills

+++ internal/character/get_hp.go
@@ -1,5 +1,5 @@
 package character
 
-func getHP(con int, siz int) int {
-	return (con + siz) / 10
+func getHP(con float32, siz float32) int {
+	return int((con + siz) / 10)
 }
